Updated Witch Vanilla Loot Table
I've made changes to the Vanilla Witch Loot Table in FMH to match the 1.20.10 Beta Preview.

I believe this is meant to be implemented, maybe at a later date in a future full release update. Based on the research done for it, I think we should get ahead of the game and implement this now.