Manifest Update
There was a lot of inconsistencies between many of the manifest.json files and World Packs...

So I incremented most of the packs, fixed some of the dependencies, made most of the packs readable, and just overall less annoying to look at.

I left bedrock Bridge alone though, as that is fine.